DrinkList now lazy loads
Might be a bit scuffed, but it works for now. It uses the name of the drink for sorting and loads 7 drinks at a time. This behavior might need to be changed later.
Also, there is logged an error when it reaches the last drink, but that's a future problem.